requirements.txt
torch
torchvision
timm
gradio
pillow
ultralytics
numpy